+++ matrix-sdk-android/src/main/java/im/vector/matrix/android/internal/session/pushers/GetPushRulesTask.kt
@@ -15,14 +15,9 @@
  */
 package im.vector.matrix.android.internal.session.pushers
 
-import com.zhuinden.monarchy.Monarchy
-import im.vector.matrix.android.api.pushrules.RuleSetKey
 import im.vector.matrix.android.api.pushrules.rest.GetPushRulesResponse
-import im.vector.matrix.android.internal.database.mapper.PushRulesMapper
-import im.vector.matrix.android.internal.database.model.PushRulesEntity
 import im.vector.matrix.android.internal.network.executeRequest
 import im.vector.matrix.android.internal.task.Task
-import im.vector.matrix.android.internal.util.awaitTransaction
 import javax.inject.Inject
 
 
@@ -30,52 +25,17 @@ internal interface GetPushRulesTask : Task<GetPushRulesTask.Params, Unit> {
     data class Params(val scope: String)
 }
 
+/**
+ * We keep this task, but it should not be used anymore, the push rules comes from the sync response
+ */
 internal class DefaultGetPushRulesTask @Inject constructor(private val pushRulesApi: PushRulesApi,
-                                                           private val monarchy: Monarchy) : GetPushRulesTask {
+                                                           private val savePushRulesTask: SavePushRulesTask) : GetPushRulesTask {
 
     override suspend fun execute(params: GetPushRulesTask.Params) {
         val response = executeRequest<GetPushRulesResponse> {
             apiCall = pushRulesApi.getAllRules()
         }
-        val scope = params.scope
-        monarchy.awaitTransaction { realm ->
-            //clear existings?
-            //TODO
-            realm.where(PushRulesEntity::class.java)
-                    .findAll()
-                    .deleteAllFromRealm()
 
-            val content = PushRulesEntity(scope).apply { kind = RuleSetKey.CONTENT }
-            response.global.content?.forEach { rule ->
-                content.pushRules.add(PushRulesMapper.map(rule))
-            }
-            realm.insertOrUpdate(content)
-
-            val override = PushRulesEntity(scope).apply { kind = RuleSetKey.OVERRIDE }
-            response.global.override?.forEach { rule ->
-                PushRulesMapper.map(rule).also {
-                    override.pushRules.add(it)
-                }
-            }
-            realm.insertOrUpdate(override)
-
-            val rooms = PushRulesEntity(scope).apply { kind = RuleSetKey.ROOM }
-            response.global.room?.forEach { rule ->
-                rooms.pushRules.add(PushRulesMapper.map(rule))
-            }
-            realm.insertOrUpdate(rooms)
-
-            val senders = PushRulesEntity(scope).apply { kind = RuleSetKey.SENDER }
-            response.global.sender?.forEach { rule ->
-                senders.pushRules.add(PushRulesMapper.map(rule))
-            }
-            realm.insertOrUpdate(senders)
-
-            val underrides = PushRulesEntity(scope).apply { kind = RuleSetKey.UNDERRIDE }
-            response.global.underride?.forEach { rule ->
-                underrides.pushRules.add(PushRulesMapper.map(rule))
-            }
-            realm.insertOrUpdate(underrides)
-        }
+        savePushRulesTask.execute(SavePushRulesTask.Params(response))
     }
 }